searxng_search/pyproject.toml
[project]
name = "searxng_search"
version = "0.1.0"
description = "Agent can help users search for information on the web."
authors = [
    { name = "Dan Higgins", email = "dphiggins@gmail.com" },
]
license = "Apache License 2.0"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk[a2a,eval]>=2.0.0a3",
    "httpx>=0.28.1",
    "litellm>=1.82.6",
    "mcp[cli]>=1.13.0",
]

